requirements.txt
Django==4.1.4
django-ckeditor==6.5.1
django-extensions==3.2.1
django-js-asset==2.0.0
python-decouple==3.6
whitenoise==6.2.0
